Extract InMemoryStateStoreTestBase
